fix: remove _cli suffix
