+++ .github/workflows/test.yaml
@@ -19,7 +19,7 @@ jobs:
             diagnostic_aggregator,
             diagnostic_common_diagnostics,
             diagnostic_updater,
-            ros2diagnostics_cli,
+            ros2diagnostics,
             self_test,
           ]
         distro: [humble, iron, rolling]
