+++ ros2diagnostics/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format2.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="2">
-  <name>ros2diagnostics_cli</name>
+  <name>ros2diagnostics</name>
   <version>1.0.0</version>
   <description>diagnostic command for ROS2 command line, parse and show diagnostics topic</description>
   <maintainer email="robo2020@gmail.com"></maintainer>
