+++ ros2diagnostics/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/ros2diagnostics
+[install]
+install_scripts=$base/lib/ros2diagnostics
