--- ros2diagnostics_cli/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/ros2diagnostics_cli
-[install]
-install_scripts=$base/lib/ros2diagnostics_cli
